go.d fix compiling for windows
