+++ src/go/logger/journal_stub.go
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+//go:build !linux
+
+package logger
+
+func isStderrConnectedToJournal() bool {
+	return false
+}
