+++ src/go/logger/logger.go
@@ -7,10 +7,7 @@ import (
 	"fmt"
 	"log/slog"
 	"os"
-	"strconv"
-	"strings"
 	"sync/atomic"
-	"syscall"
 
 	"github.com/netdata/netdata/go/plugins/pkg/executable"
 
@@ -81,24 +78,3 @@ func (l *Logger) mute(v bool) {
 func (l *Logger) isNil() bool { return l == nil || l.sl == nil }
 
 var nilLogger = New()
-
-func isStderrConnectedToJournal() bool {
-	stream := os.Getenv("JOURNAL_STREAM")
-	if stream == "" {
-		return false
-	}
-
-	idx := strings.IndexByte(stream, ':')
-	if idx <= 0 {
-		return false
-	}
-
-	dev, ino := stream[:idx], stream[idx+1:]
-
-	var stat syscall.Stat_t
-	if err := syscall.Fstat(int(os.Stderr.Fd()), &stat); err != nil {
-		return false
-	}
-
-	return dev == strconv.Itoa(int(stat.Dev)) && ino == strconv.FormatUint(stat.Ino, 10)
-}
